+++ src/cli/config.rs
@@ -53,13 +53,16 @@ pub enum ConfigError {
 pub struct CliConfig {
     /// Path to the configuration file (not serialized)
     #[serde(skip)]
-    path:     PathBuf,
+    path:                PathBuf,
     /// Groups of repositories and relays.
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub sets: Vec<RepoRelaySet>,
+    pub sets:            Vec<RepoRelaySet>,
     /// The default PoW difficulty
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub pow:  Option<u8>,
+    pub pow:             Option<u8>,
+    /// List of fallback relays used if no fallback relays was provided.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub fallback_relays: Option<Vec<RelayUrl>>,
 }
 
 /// A named group of repositories and relays.